refactor: 불필요한 주석 삭제 및 JwtTokenProvider service -> infrastructure 이동
